ref(server): Split envelope module into multiple smaller ones
